fix bigs in import view

+++ settings/models.py
@@ -28,8 +28,8 @@ class CountryConfig(models.Model):
 
 class LevelConfig(models.Model):
     country=models.ForeignKey(CountryConfig,on_delete=models.CASCADE,blank=True,null=True)
-    maxpop = models.DecimalField(max_digits=10, decimal_places=2, blank=True, null=True)
-    minpop = models.DecimalField(max_digits=10, decimal_places=2, blank=True, null=True)
+    maxpop = models.IntegerField( blank=True, null=True)
+    minpop = models.IntegerField( blank=True, null=True)
     uppervol = models.DecimalField(max_digits=10, decimal_places=2, blank=True, null=True)
     undervol = models.DecimalField(max_digits=10, decimal_places=2, blank=True, null=True)
     m25vol = models.DecimalField(max_digits=10, decimal_places=2, blank=True, null=True)
